+++ NEWS.md
@@ -1,7 +1,13 @@
-# rbacon 3.3.2
-* radiocarbon ages can now be entered as F14C or pMC values (e.g. for postbomb dates). Please indicate using the F14C or pMC option which ones are in F14C/pMC (e.g., Bacon(F14C=2:4) if dates 2 to 4 are in F14C). These values will then automatically be rewritten as C14 ages.
+# rbacon 3.4.0
+* radiocarbon ages can now be entered as F14C or pMC values (e.g. for postbomb dates). Please indicate using the F14C or pMC option which ones are in F14C/pMC (e.g., `Bacon(F14C=2:4)` if dates 2 to 4 are in F14C). These values will then automatically be rewritten as C14 ages.
 * the proxy.ghost function now invisibly returns a table of the values used in the grid composing the greyscale plot.
-* add.dates now plots additional dates as expected, also when rotate.axes=TRUE.
+* add.dates now plots additional dates as expected, also when rotate.axes=TRUE (although rotate.axes still has to be set to TRUE after using the agedepth function with agedepth=TRUE).
+* the agedepth function now treats provided values for d.min and d.max better.
+* greyscale 'ghost' plots (`agedepth`, `accrate.depth.ghost`, `accrate.age.ghost`, `proxy.ghost`) should now plot with fewer disturbances such as lines. If this doesn't work, please use the option 'use.raster=FALSE'.
+* if available on your system, 'cairo_pdf' will be used to plot pdfs.
+* after a run, the posteriors (for accumulation rate, memory, and if present hiatus, phi and supported) are now summarized in a message.
+* The heights of the distributions of the dates can now also be set through a variable `ex`, which could either be of length 1, or have a value for each date in the core. This way, some dates can be plotted at different heights. 
+* When running Bacon as `tmp <- Bacon(save.info=FALSE)`, no additional variables beside 'tmp' will be saved in the session. For subsequent calculations, provide 'tmp', e.g. as in `Bacon.hist(20:40, set=tmp)`. The default remains to save an object 'info' to the session, and this object will then be used to make any further calculations using `set=get('info')`. 
 
 # rbacon 3.3.1
 * now uses the updated rice package (which loads the data rintcal package).